+++ lib/plugins/aws/package/compile/events/schedule.js
