+++ lib/plugins/index.js
@@ -34,7 +34,7 @@ module.exports = [
   require('./aws/rollbackFunction/index.js'),
   require('./aws/package/compile/layers/index.js'),
   require('./aws/package/compile/functions/index.js'),
-  require('./aws/package/compile/events/schedule/index.js'),
+  require('./aws/package/compile/events/schedule.js'),
   require('./aws/package/compile/events/s3/index.js'),
   require('./aws/package/compile/events/apiGateway/index.js'),
   require('./aws/package/compile/events/websockets/index.js'),
