+++ test/unit/lib/plugins/aws/package/compile/events/schedule.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsCompileScheduledEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/schedule/index');
-const Serverless = require('../../../../../../../../../lib/Serverless');
+const AwsProvider = require('../../../../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsCompileScheduledEvents = require('../../../../../../../../lib/plugins/aws/package/compile/events/schedule');
+const Serverless = require('../../../../../../../../lib/Serverless');
 
 describe('AwsCompileScheduledEvents', () => {
   let serverless;
